+++ src/main/java/com/bokmcdok/butterflies/client/model/ButterflyModel.java
@@ -1,7 +1,7 @@
-package com.bokmcdok.butterflies.model;
+package com.bokmcdok.butterflies.client.model;
 
 import com.bokmcdok.butterflies.ButterfliesMod;
-import com.bokmcdok.butterflies.world.entity.ambient.Butterfly;
+import com.bokmcdok.butterflies.world.entity.animal.Butterfly;
 import net.minecraft.client.model.HierarchicalModel;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.ModelPart;
@@ -13,11 +13,14 @@ import net.minecraft.client.model.geom.builders.MeshDefinition;
 import net.minecraft.client.model.geom.builders.PartDefinition;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * A model of a butterfly.
  */
+@OnlyIn(Dist.CLIENT)
 public class ButterflyModel  extends HierarchicalModel<Butterfly> {
 
     //  Holds the layers for the butterfly.