+++ src/main/java/com/bokmcdok/butterflies/registries/ItemRegistry.java
@@ -1,8 +1,8 @@
 package com.bokmcdok.butterflies.registries;
 
 import com.bokmcdok.butterflies.ButterfliesMod;
-import com.bokmcdok.butterflies.world.entity.ambient.Butterfly;
-import com.bokmcdok.butterflies.world.entity.ambient.Caterpillar;
+import com.bokmcdok.butterflies.world.entity.animal.Butterfly;
+import com.bokmcdok.butterflies.world.entity.animal.Caterpillar;
 import com.bokmcdok.butterflies.world.item.BottledButterflyItem;
 import com.bokmcdok.butterflies.world.item.ButterflyEggItem;
 import com.bokmcdok.butterflies.world.item.ButterflyNetItem;
@@ -35,8 +35,8 @@ public class ItemRegistry {
             () -> new BottledButterflyItem(new Item.Properties().stacksTo(1)));
 
     // Butterfly Scroll
-    public static final RegistryObject<Item> BUTTERFLY_SCROLL = INSTANCE.register(ButterflyScrollItem.NAME,
-            () -> new ButterflyScrollItem(new Item.Properties()));
+    public static final RegistryObject<Item> BUTTERFLY_SCROLL =
+            INSTANCE.register(ButterflyScrollItem.NAME, ButterflyScrollItem::new);
 
     // Butterfly Eggs - Eggs that will eventually hatch into a caterpillar.
     public static final RegistryObject<Item> ADMIRAL_BUTTERFLY_EGG = INSTANCE.register(ButterflyEggItem.ADMIRAL_NAME,