+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## Changelog
 
+### 1.1.1  (2023-11-12)
+-   Butterfly scrolls can now be placed around the world.
+
 ### 1.1.0  (2023-11-05)
--   Added butterfly scrolls
+-   Added butterfly scrolls.
+
+### 1.0.2 (2023-11-06)
+-   Added names to the entities to better support modpacks.
 
 ### 1.0.1 (2023-10-29)
 -   Updated Forge to current recommended version (47.2.0).